requirements.txt
Flask==2.3.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0